+++ packages/spacecat-shared-data-access/src/models/suggestion/suggestion.model.js
@@ -21,6 +21,8 @@ import BaseModel from '../base/base.model.js';
  * @extends BaseModel
  */
 class Suggestion extends BaseModel {
+  static ENTITY_NAME = 'Suggestion';
+
   static STATUSES = {
     NEW: 'NEW',
     APPROVED: 'APPROVED',
@@ -41,7 +43,7 @@ class Suggestion extends BaseModel {
     CONFIG_UPDATE: 'CONFIG_UPDATE',
   };
 
-  // add your customized methods here
+  // add your customized method  here
 }
 
 export default Suggestion;