+++ packages/spacecat-shared-data-access/test/it/site/site.test.js
@@ -50,7 +50,8 @@ describe('Site IT', async () => {
   let sampleData;
   let Site;
 
-  before(async () => {
+  before(async function () {
+    this.timeout(10000);
     sampleData = await seedDatabase();
 
     const dataAccess = getDataAccess();