+++ packages/spacecat-shared-data-access/test/it/util/db.js
@@ -51,6 +51,7 @@ export const TEST_DA_CONFIG = {
   tableNameSiteTopPages: 'spacecat-services-site-top-pages',
   tableNameSites: 'spacecat-services-sites',
   tableNamePageIntents: 'spacecat-services-page-intents',
+  tableNamePageCitabilities: 'spacecat-services-page-citabilities',
   tableNameSpacecatData: 'spacecat-data',
 };
 
@@ -75,15 +76,9 @@ const getDynamoClients = (config = {}) => {
   return { dbClient, docClient };
 };
 
-// Minimal logger that doesn't output debug messages to prevent memory issues
-const minimalLogger = {
-  info: () => {},
-  warn: () => {},
-  error: () => {},
-  debug: () => {}, // Disable debug to prevent massive ElectroDB logs
-};
-
-export const getDataAccess = (config, logger = minimalLogger) => {
+export const getDataAccess = (config, logger = console) => {
+  // eslint-disable-next-line no-param-reassign
+  logger.debug = () => {};
   const { dbClient } = getDynamoClients(config);
   return createDataAccess(TEST_DA_CONFIG, logger, dbClient);
 };